+++ anyway/widgets/suburban_widgets/AccidentCountByAccidentTypeWidget.py
