+++ anyway/widgets/suburban_widgets/AccidentCountByDayNightWidget.py
@@ -1,0 +1,24 @@
+from anyway.RequestParams import RequestParams
+from anyway.infographics_utils import register, get_accidents_stats
+from anyway.models import AccidentMarkerView
+from anyway.widgets.suburban_widgets.SubUrbanWidget import SubUrbanWidget
+
+
+@register
+class AccidentCountByDayNightWidget(SubUrbanWidget):
+    name: str = "accident_count_by_day_night"
+
+    def __init__(self, request_params: RequestParams):
+        super().__init__(request_params, type(self).name)
+        self.rank = 10
+        self.text = {"title": "כמות תאונות ביום ובלילה"}
+
+    def generate_items(self) -> None:
+        self.items = get_accidents_stats(
+            table_obj=AccidentMarkerView,
+            filters=self.request_params.location_info,
+            group_by="day_night_hebrew",
+            count="day_night_hebrew",
+            start_time=self.request_params.start_time,
+            end_time=self.request_params.end_time,
+        )
