+++ anyway/widgets/suburban_widgets/AccidentCountByHourWidget.py
@@ -1,0 +1,24 @@
+from anyway.RequestParams import RequestParams
+from anyway.infographics_utils import register, get_accidents_stats
+from anyway.models import AccidentMarkerView
+from anyway.widgets.suburban_widgets.SubUrbanWidget import SubUrbanWidget
+
+
+@register
+class AccidentCountByHourWidget(SubUrbanWidget):
+    name: str = "accident_count_by_hour"
+
+    def __init__(self, request_params: RequestParams):
+        super().__init__(request_params, type(self).name)
+        self.rank = 11
+        self.text = {"title": "כמות תאונות לפי שעה"}
+
+    def generate_items(self) -> None:
+        self.items = get_accidents_stats(
+            table_obj=AccidentMarkerView,
+            filters=self.request_params.location_info,
+            group_by="accident_hour",
+            count="accident_hour",
+            start_time=self.request_params.start_time,
+            end_time=self.request_params.end_time,
+        )
