+++ anyway/widgets/suburban_widgets/AccidentCountByRoadLightWidget.py
@@ -1,0 +1,24 @@
+from anyway.RequestParams import RequestParams
+from anyway.infographics_utils import register, get_accidents_stats
+from anyway.models import AccidentMarkerView
+from anyway.widgets.suburban_widgets.SubUrbanWidget import SubUrbanWidget
+
+
+@register
+class AccidentCountByRoadLightWidget(SubUrbanWidget):
+    name: str = "accident_count_by_road_light"
+
+    def __init__(self, request_params: RequestParams):
+        super().__init__(request_params, type(self).name)
+        self.rank = 12
+        self.text = {"title": "כמות תאונות לפי תאורה"}
+
+    def generate_items(self) -> None:
+        self.items = get_accidents_stats(
+            table_obj=AccidentMarkerView,
+            filters=self.request_params.location_info,
+            group_by="road_light_hebrew",
+            count="road_light_hebrew",
+            start_time=self.request_params.start_time,
+            end_time=self.request_params.end_time,
+        )
