+++ anyway/widgets/suburban_widgets/HeadOnCollisionsComparisonWidget.py
