+++ anyway/widgets/suburban_widgets/MostSevereAccidentsTableWidget.py
