+++ anyway/widgets/suburban_widgets/PedestrianInjuredInJunctionsWidget.py
@@ -1,0 +1,24 @@
+from anyway.RequestParams import RequestParams
+from anyway.infographics_utils import register
+from anyway.widgets.suburban_widgets.SubUrbanWidget import SubUrbanWidget
+
+
+@register
+class PedestrianInjuredInJunctionsWidget(SubUrbanWidget):
+    name: str = "pedestrian_injured_in_junctions"
+
+    def __init__(self, request_params: RequestParams):
+        super().__init__(request_params, type(self).name)
+        self.rank = 23
+        self.text = {"title": "מספר נפגעים הולכי רגל בצמתים - רחוב בן יהודה, תל אביב"}
+
+    def generate_items(self) -> None:
+        self.items = PedestrianInjuredInJunctionsWidget.pedestrian_injured_in_junctions_mock_data()
+
+    @staticmethod
+    def pedestrian_injured_in_junctions_mock_data():  # Temporary for Frontend
+        return [
+            {"street name": "גורדון י ל", "count": 18},
+            {"street name": "אידלסון אברהם", "count": 10},
+            {"street name": "פרישמן", "count": 7},
+        ]
